+++ backend/src/main/java/de/mightypc/backend/model/specs/createspecs/CreateGpu.java
@@ -1,9 +1,11 @@
 package de.mightypc.backend.model.specs.createspecs;
 
+import java.math.BigDecimal;
+
 public record CreateGpu (
         String name,
         String description,
-        float price,
+        BigDecimal price,
         int performance,
         int energyConsumption,
         float rating
