+++ backend/src/main/java/de/mightypc/backend/model/specs/createspecs/CreateHdd.java
@@ -1,11 +1,13 @@
 package de.mightypc.backend.model.specs.createspecs;
 
+import java.math.BigDecimal;
+
 public record CreateHdd(
         String name,
         String description,
         int capacity,
         int energyConsumption,
-        float price,
+        BigDecimal price,
         float rating
 ) {
 }
