+++ backend/src/main/java/de/mightypc/backend/model/specs/createspecs/CreateMotherboard.java
@@ -3,10 +3,12 @@ package de.mightypc.backend.model.specs.createspecs;
 import de.mightypc.backend.model.specs.CPU;
 import de.mightypc.backend.model.specs.GPU;
 
+import java.math.BigDecimal;
+
 public record CreateMotherboard(
         String name,
         String description,
-        int price,
+        BigDecimal price,
         int energyConsumption,
         GPU[] graphicCardCompatibility,
         CPU[] processorCompatibility,
